requirements.txt
python-telegram-bot==20.6
psycopg2-binary==2.9.9
openai==1.50.0
python-dotenv==1.0.1
matplotlib==3.9.2
